+++ internal/mode/usage/markdown/index.go
@@ -2,7 +2,6 @@ package markdown
 
 import (
 	"fmt"
-	"strings"
 
 	"github.com/dogmatiq/ferrite/variable"
 )
@@ -11,41 +10,34 @@ func (r *renderer) renderIndex() {
 	r.line("## Index")
 	r.gap()
 
-	for _, v := range r.Specs {
-		r.renderIndexItem(v)
-	}
-}
+	var t table
 
-func (r *renderer) renderIndexItem(s variable.Spec) {
-	var w strings.Builder
+	t.AddRow("Name", "Optionality", "Description")
 
-	style := ""
-	label := ""
-	if s.IsDeprecated() {
-		style = "~~"
-		label = "deprecated"
-	} else if def, ok := s.Default(); ok {
-		label = fmt.Sprintf("defaults to `%s`", def.Quote())
-	} else if !s.IsRequired() {
-		label = "optional"
-	} else if len(variable.Relationships[variable.DependsOn](s)) != 0 {
-		label = "conditional"
-	} else {
-		style = "**"
-		label = "required"
+	for _, s := range r.Specs {
+		name := r.linkToSpec(s)
+		optionality := "required"
+
+		if s.IsDeprecated() {
+			name = "~~" + name + "~~"
+			optionality = "optional, deprecated"
+		} else if def, ok := s.Default(); ok {
+			optionality = fmt.Sprintf("defaults to `%s`", def.Quote())
+		} else if !s.IsRequired() {
+			optionality = "optional"
+		} else if len(variable.Relationships[variable.DependsOn](s)) != 0 {
+			optionality = "conditional"
+		}
+
+		t.AddRow(name, optionality, s.Description())
 	}
 
-	w.WriteString("- ")
-	w.WriteString(style)
-	w.WriteString(r.linkToSpec(s))
-	w.WriteString(style)
-	w.WriteString(" — ")
-	w.WriteString(style)
-	w.WriteString(s.Description())
-	w.WriteString(style)
-	w.WriteString(" (")
-	w.WriteString(label)
-	w.WriteString(")")
-
-	r.line(w.String())
+	t.WriteTo(r.Output)
 }
+
+// | Name                 | Optionality           | Description                            |
+// | -------------------- | --------------------- | -------------------------------------- |
+// | ~~[`BIND_ADDRESS`]~~ | optional, deprecated  | listen address for the HTTP server |
+// | [`BIND_HOST`]        | defaults to `0.0.0.0` | listen host for the HTTP server        |
+// | [`BIND_PORT`]        | defaults to `8080`    | listen port for the HTTP server        |
+// | [`BIND_VERSION`]     | defaults to `4`       | IP version for the HTTP server         |